fixed some minors changes

+++ bluememo/src/main/java/com/bluedigi/bluememo/channel/telegram/outbound/infrastructure/api/dto/response/SendMessageResponse.java
@@ -8,7 +8,7 @@ public record SendMessageResponse(
         @JsonProperty("message_id")
         Integer messageId,
 
-        Integer date,
+        Long date,
 
         TelegramChatResponse chat,
